Fix some VS2015 incompatibilities in refinery.
VS2015 doesn't like some of the implicit casts and hash_[map|set] is gone !

R=manzagop@chromium.org

Review URL: https://codereview.chromium.org/1847203002 .

+++ syzygy/refinery/types/type_repository.cc
@@ -38,7 +38,7 @@ TypePtr TypeRepository::GetType(TypeId id) const {
 
 TypeId TypeRepository::AddType(TypePtr type) {
   DCHECK(type);
-  TypeId id = types_.size() + 1;
+  TypeId id = static_cast<uint16_t>(types_.size() + 1);
 
   bool result = AddTypeWithId(type, id);
